Fix schedule create notification review issues
